+++ platform/linux/opengl.c
@@ -20,6 +20,7 @@
 #include <EGL/egl.h>
 #include <EGL/eglext.h>
 #include <EGL/eglplatform.h>
+#include <GL/gl.h>
 
 #define MODULE_NAME "opengl"
 
@@ -115,6 +116,16 @@ struct p_opengl_ctx * p_opengl_create_context(struct p_window *win)
 
     if (load_opengl_functions(&ctx->functions, ctx->egl.fns.eglGetProcAddress))
         goto_error("Failed to load OpenGL functions.");
+
+    /* Render 1 empty frame */
+    ctx->functions.glClearColor(0.0f, 0.0f, 0.0f, 0.0f);
+    ctx->functions.glClear(GL_COLOR_BUFFER_BIT);
+    p_opengl_swap_buffers(ctx, win);
+    if (p_window_swap_buffers(win, P_WINDOW_PRESENT_NOW)
+         == P_WINDOW_SWAP_BUFFERS_FAIL)
+    {
+        goto_error("Failed to present the first frame");
+    }
 
     return ctx;